+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Config to disable material mixing in the crafting grid when crafting parts. This is left enabled (old behavior by default).
 ### Fixed
 - Elytra not working on back slot with Caelus API [#302]
+- Armor plates displaying the wrong armor value
 
 ## [2.5.2] - 2021-01-17
 ### Fixed