+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.5.3] - 2021-01-24
 ### Added
 - Fragments can now be used to fill repair kits (1/8 of a unit per) [#311]
 - Add netherwood wood (log with the side texture on all side), with stripped variant [#301]